test full run of demo scripts

+++ bin/destroy.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 set -euo pipefail
 
-NAMESPACE=$(jq -r .namespace < "${DIRNAME}/../config.json")
+DIRNAME=$(dirname "$0")
 
-kubectl delete ns "${NAMESPACE}"
+NAMESPACE=$(jq -r .namespace < "${DIRNAME}/../config.json")
+if kubectl get ns "${NAMESPACE}" >/dev/null; then
+    echo -n "This will namespace ${NAMESPACE}, please type 'yes' if you sure: "
+    read -r answer
+    [ "$answer" == "yes" ] && kubectl delete ns "${NAMESPACE}"
+    exit 0
+fi
